# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "complain_management_sys",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://127.0.0.1"
    }
  ],
  "paths": {
...
    "/user/update/11": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "update_user",
        "description": "update_user",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "example": "admin"
                  },
                  "role_type": {
                    "type": "string",
                    "example": "admin"
                  }
                }
              },
              "examples": {
                "update_user": {
                  "value": {
                    "password": "admin",
                    "role_type": "admin"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New Request",
            "headers": {
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "close"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "string",
                  "example": "97"
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Fri, 08 Dec 2023 19:46:40 GMT"
                }
              },
              "Server": {
                "schema": {
                  "type": "string",
                  "example": "Werkzeug/3.0.1 Python/3.11.5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "errors": {
                      "type": "array",
                      "items": {},
                      "example": []
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "status_code": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "New Request": {
                    "value": {
                      "data": [],
                      "errors": [],
                      "message": "User updated successfully",
                      "status_code": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}